chore: Support a `x-cdn-url` header to set the base for assets
BREAKING CHANGE: With earlier versions of this module the cdnUrl should have pointed to the  `/assets` directory/path.
We've changed the behavior of that to make it more transparent. Please reconfigure the cdnUrl to point to the root of your server instead of the `/assets` directory as we append that automatically.